Adding Loan entity and its controller, repository and service
Add validation of the id when updating and deleting a register in both the Book entity and the Loan entity

+++ src/main/java/com/crud/crudbook/service/BookService.java
@@ -2,11 +2,10 @@ package com.crud.crudbook.service;
 
 import com.crud.crudbook.entity.Book;
 import com.crud.crudbook.repository.BookRepository;
-import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BookService implements BookServiceInterface{
@@ -26,11 +25,20 @@ public class BookService implements BookServiceInterface{
 
     @Override
     public Book updateBook(Book book) {
+        Optional<Book> bookExist = bookRepository.findById(book.getId());
+        if(bookExist.isEmpty()){
+            return null;
+        }
         return bookRepository.save(book);
     }
 
     @Override
-    public void deleteBook(Long id) {
-        bookRepository.deleteById(id);
+    public Optional<Book> deleteBook(Long id) {
+        Optional<Book> bookExist = bookRepository.findById(id);
+        if(bookExist.isPresent()){
+            bookRepository.deleteById(id);
+            return bookExist;
+        }
+        return bookExist;
     }
 }
